backend/requirements.txt
# ProteinHub Requirements
# 后端依赖

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
PyJWT==2.8.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0

# 测试依赖
pytest==7.4.3
pytest-flask==1.3.0
